organization of code continued
