+++ tomcat_speech/evaluation_scripts/evaluate_single_task.py
@@ -11,8 +11,8 @@ import torch
 import numpy as np
 import random
 
-from tomcat_speech.train_and_test_models.train_and_test_models import evaluate
-from tomcat_speech.train_and_test_models.train_and_test_utils import make_train_state
+from tomcat_speech.training_and_evaluation_functions.train_and_test_models import evaluate
+from tomcat_speech.training_and_evaluation_functions.train_and_test_utils import make_train_state
 from tomcat_speech.models.multimodal_models import MultitaskModel
 
 # import MultitaskObject and Glove from preprocessing code
