+++ tomcat_speech/training_and_evaluation_functions/asist_analysis_functions.py
@@ -6,7 +6,7 @@ import torch
 import sys
 
 from tomcat_speech.data_prep.asist_data.asist_dataset_creation import AsistDataset
-from tomcat_speech.train_and_test_models.train_and_test_without_gold_labels import (
+from tomcat_speech.training_and_evaluation_functions.train_and_test_without_gold_labels import (
     multitask_predict_without_gold_labels
 )
 from tomcat_speech.models.multimodal_models import MultitaskModel
