+++ tomcat_speech/training_and_evaluation_functions/train_and_test_personality_multitask.py
@@ -4,7 +4,7 @@ from sklearn.metrics import confusion_matrix
 from sklearn.metrics import classification_report
 from sklearn.metrics import precision_recall_fscore_support
 
-from tomcat_speech.train_and_test_models.train_and_test_utils import update_train_state
+from tomcat_speech.training_and_evaluation_functions.train_and_test_utils import update_train_state
 
 
 def personality_as_multitask_train_and_predict(
