+++ tomcat_speech/training_and_evaluation_functions/train_and_test_without_gold_labels.py
