+++ tomcat_speech/training_scripts/train_asist.py
@@ -4,10 +4,10 @@
 #       if your acoustic features have not been extracted from audio
 
 from tomcat_speech.data_prep.asist_data.asist_dataset_creation import AsistDataset
-from tomcat_speech.train_and_test_models.train_and_test_models import *
+from tomcat_speech.training_and_evaluation_functions.train_and_test_models import *
 from tomcat_speech.models.plot_training import *
 from tomcat_speech.models.multimodal_models import *
-from tomcat_speech.train_and_test_models.train_and_test_utils import make_train_state
+from tomcat_speech.training_and_evaluation_functions.train_and_test_utils import make_train_state
 
 sys.path.append("../multimodal_data_preprocessing")
 from utils.data_prep_helpers import (
