+++ tomcat_speech/training_scripts/train_multitask.py
@@ -10,9 +10,9 @@ from datetime import date
 
 from tomcat_speech.data_prep.samplers import RandomOversampler
 # from tomcat_speech.data_prep.samplers import RandomOversampler
-from tomcat_speech.train_and_test_models.train_and_test_models import train_and_predict
+from tomcat_speech.training_and_evaluation_functions.train_and_test_models import train_and_predict
 from tomcat_speech.models.plot_training import *
-from tomcat_speech.train_and_test_models.train_and_test_utils import (
+from tomcat_speech.training_and_evaluation_functions.train_and_test_utils import (
     set_cuda_and_seeds,
     select_model,
     make_train_state
